refactor(mcp): move 4 toolService.mjs files back to MCP server directories (#11107)
Per #11107 retrospective architectural fix: 4 toolService.mjs files were moved
from MCP server dirs to SDK during M6 mechanical migration. These files are
pure MCP-server-suitable glue (ToolService instantiation + openapi.yaml-bound
serviceMapping + per-server callTool/listTools dispatch) — they don't satisfy
the SDK-suitable test (cross-server reusable, independent of MCP framing).

Operator-surfaced anchor from #11107 body:

> "the goal was to move services outside mcp servers into the sdk. you guys
>  rubber stamped it, and literally moved all one by one. for most services
>  correct, except for the toolService files (one per server). these literally
>  just map service methods to mcp tools. and these REALLY belong into the
>  mcp servers."

This PR moves all 4 back to flat layout (`ai/mcp/server/<name>/toolService.mjs`).

Files moved:
- ai/services/knowledge-base/toolService.mjs → ai/mcp/server/knowledge-base/toolService.mjs
- ai/services/memory-core/toolService.mjs → ai/mcp/server/memory-core/toolService.mjs
- ai/services/neural-link/toolService.mjs → ai/mcp/server/neural-link/toolService.mjs
- ai/services/github-workflow/toolService.mjs → ai/mcp/server/github-workflow/toolService.mjs

Per-file path adjustments:
- ToolService import: '../../mcp/ToolService.mjs' → '../ToolService.mjs' (one less level)
- openapi.yaml path: '../../mcp/server/<name>/openapi.yaml' → 'openapi.yaml' (co-resident)
- Service-class imports: './<Service>.mjs' → '../../../services/<name>/<Service>.mjs' (reach into SDK for reusable logic)
- (neural-link only) getCurrentTurnId import: '../../mcp/server/neural-link/Server.mjs' → './Server.mjs' (sibling)
- (github-workflow only) config import: '../../mcp/server/github-workflow/config.mjs' → './config.mjs' (sibling)

Server.mjs import per server:
- '../../../services/<name>/toolService.mjs' → './toolService.mjs' (sibling)

Test file path updates (5 files):
- test/playwright/unit/ai/services/knowledge-base/KBRecorderService.spec.mjs
- test/playwright/unit/ai/services/github-workflow/toolService.spec.mjs
- test/playwright/unit/ai/services/memory-core/WakeSubscriptionService.spec.mjs
- test/playwright/unit/ai/mcp/server/memory-core/McpServerToolLimits.spec.mjs
- test/playwright/mcp/github-workflow/ToolRegistration.spec.mjs

File-system server scope note: `ai/mcp/server/file-system/services/toolService.mjs`
remains at its services/ subfolder location (skipped from M6 mechanical migration,
hence skipped from this back-migration too). Normalization to flat layout for
file-system server is a separate scope concern (could be follow-up if substrate
coherence requires it).

13 files changed, 53 insertions/53 deletions (1:1 path rewrites; no logic changes).
