+++ ai/mcp/server/memory-core/toolService.mjs
@@ -1,22 +1,20 @@
 import path                     from 'path';
 import {fileURLToPath}          from 'url';
-import ToolService              from '../../mcp/ToolService.mjs';
-import {
-    Memory_DatabaseService          as DatabaseService,
-    Memory_ChromaLifecycleService   as ChromaLifecycleService,
-    Memory_GraphService             as GraphService,
-    Memory_HealthService            as HealthService,
-    Memory_Service                  as MemoryService,
-    Memory_SessionService           as SessionService,
-    Memory_SummaryService           as SummaryService,
-    Memory_MailboxService           as MailboxService,
-    Memory_PermissionService        as PermissionService,
-    Memory_WakeSubscriptionService  as WakeSubscriptionService
-} from '../../services.mjs';
+import ToolService              from '../../ToolService.mjs';
+import DatabaseService          from '../../../services/memory-core/DatabaseService.mjs';
+import ChromaLifecycleService   from '../../../services/memory-core/lifecycle/ChromaLifecycleService.mjs';
+import GraphService             from '../../../services/memory-core/GraphService.mjs';
+import HealthService            from '../../../services/memory-core/HealthService.mjs';
+import MemoryService            from '../../../services/memory-core/MemoryService.mjs';
+import SessionService           from '../../../services/memory-core/SessionService.mjs';
+import SummaryService           from '../../../services/memory-core/SummaryService.mjs';
+import MailboxService           from '../../../services/memory-core/MailboxService.mjs';
+import PermissionService        from '../../../services/memory-core/PermissionService.mjs';
+import WakeSubscriptionService  from '../../../services/memory-core/WakeSubscriptionService.mjs';
 
 const __filename      = fileURLToPath(import.meta.url);
 const __dirname       = path.dirname(__filename);
-const openApiFilePath = path.join(__dirname, '../../mcp/server/memory-core/openapi.yaml');
+const openApiFilePath = path.join(__dirname, 'openapi.yaml');
 
 const serviceMapping = {
     add_memory            : MemoryService           .addMemory           .bind(MemoryService),
